Expose more properties on KuduScanner
Also move some initialization up to KuduScanner
